7_Demo/digital_twin/requirements.txt
numpy>=1.24
flask>=3.0
flask-socketio>=5.3
pyserial>=3.5
